Refactor pool persistence in node manager and add command to list the full ledger of a staking pool
The current change refactors the way pools are tracked in the node manager. A more straightforward version, PersistedPoolInfo, is introduced to better track pool usage by the local node. This makes the process more efficient by preventing confusion about staker/staked number mismatches between the local config file and on-chain data.
A command was also added to show the full staking ledger for a given pool.