Add node assignment logic and refactor reti daemon to no longer use a disk based config but always fetch entirely from the chain.
The validator contract now handles adding staking pools to specific nodes as well as allowing pools to be moved to different nodes. Node assignment logic was updated to allow the addition of staking pools to specific nodes. The Overall flow of the code was enhanced without affecting the daemon functionality. Changes also included some minor Fixes and improvements to code readability.