refactor(ui): add stake and unstake modals
This removes the trigger from inside the modal components so any trigger can be used.

Now there is only one modal instance per page, where the validator data they use are stored in a state variable. Setting the variable to either a validator or null determines whether they are open/closed.

With this approach, triggers can now live anywhere: as a button in a table tow, or a dropdown menu item. The trigger sets a new validator to the state variable which opens and initializes the modal with the relevant data.