refactor(ui): properties match contracts
Naming of properties in the UI was previously not consistent with the contracts, usually for the sake of brevity but it made the logic harder to parse/debug.

Now the properties match the wording exactly, but are converted to camelCase (vs PascalCase)

Also, rather than merge the properties of ValidatorConfig and ValidatorState into a single object, the Validator type now has an `id` property (plucked from config), `config`, and `state` properties, where they are separated.

All BigInts are coerced into numbers except ALGO/token amounts, which remain BigInt.