fix(ui): update validator config/state types
uint16 numbers are expected to be typed as `number` not `bigint`, but they are currently returned by the contract clients as BigInts. At runtime they need to be coerced to a Number.

Treating this as a bug, they are now typed as `number | bigint`. This way they are still coerced, and if the bug is fixed it would be non-breaking.