Update Node Manager and Algorand smart contract bootstrap logic
This commit enhances the Node Manager by adding the ability to specify environment files via environment variables for configuration. It also updates the Algorand smart contract bootstrap logic to be able to deploy to different networks and adds associated dependencies.
Also includes new rules to .dockerignore and .gitignore, a new Makefile for Docker building and pushing, and modifies the Dockerfile-nodemgr.