Update value retrieval and improve functionality in nodemgr
This commit refactors the value retrieval logic in the nodemgr section, replacing the use of `.Value` with direct type-casting methods such as `.String`, `.Uint` or `.Bool`. It also improves functionality by allowing the option to use the hostname as the node number when in a Kubernetes environment. Additionally, mnemonics are now loaded from environmental variables simply containing "_MNEMONIC" for compatibility w/ algokit conventions.