feat(contracts): Update epoch reward calculation so that Validator has more data for future arc28 events and remove logging in staking contracts
The changes update the functions responsible for reward calculation and logging in the staking contracts. More parameters are included in the `stakeUpdatedViaRewards` function such as validator commission and the amount sent to the fee sink in case of pool saturation. The logs during staking reward distribution have also been removed as this data will ultimately be reported as arc28 events via the call from pool to validator registry on each epoch.