Branch cleanup, merging dev-archive in preparation for clean 'dev'
* chore(ui): add testnet Reti app ID to .env.template

* chore(docs): Update root readme to link to gitbook maintained docs/ direcotry

* chore: remove autosave settings from workspace root

* fix(ui): don't convert algod port to a number

* feat(ui): add knight logo for both dark/light modes as new banner logo.

* Move logos to public/img
Remove unused assets/logo.svg

* feat(ui): better handling of protected/unprotected routes (#2)

* feat(ui): dashboard and details routes no longer protected

By using the fee-sink as the sender address, simulate contract calls do not require the user to be connected.

Instead of showing a connect wallet menu on the root index route and redirecting to /dashboard, now the index route is the dashboard.

* feat(ui): use router context to protect routes

* fix(ui): disable actions that require connected wallet

* fix(daemon): Update key checking and going online error message daemon

Error message in daemon for failure to go online wasn't including actual node error.
The code to determine when to go online against a 'newer' key wasn't using correct check on determining when key was in valid range.  It now uses 'VoteFirstValid' for comparison instead of 'EffectiveFirstValid'.  Keys not yet active have no EffectiveFirstValid value.

* feat(contracts): Update epoch reward calculation so that Validator has more data for future arc28 events and remove logging in staking contracts

The changes update the functions responsible for reward calculation and logging in the staking contracts. More parameters are included in the `stakeUpdatedViaRewards` function such as validator commission and the amount sent to the fee sink in case of pool saturation. The logs during staking reward distribution have also been removed as this data will ultimately be reported as arc28 events via the call from pool to validator registry on each epoch.