feat(ui): use default form mode for AddValidatorForm
Using `onBlur` mode is not the best UX and triggers unnecessary renders. It also has the potential to reset manually-triggered validation errors for the `nfdForInfo` field.

This commit changes the form mode to `onSubmit` which is the default mode for the `react-hook-form` library. The user will only see validation errors when they attempt to submit the form.