fix(ui): improved asset gating logic
Fetches all assets from creator address instead of checking the `creator` property of the user's individual assets (each requiring a separate request). This should improve performance and dramatically reduce the number of requests made to the node.

The gating asset IDs are determined in the initial fetch of validator data and added to the `gatedAssets` property in the validator data object.