feat(contracts): No longer assert if there is nothing to be paid out. This is so last payout and epoch number are both updated every epoch regardless if anything was paid out.
This will ensure the time in epoch is always correct even if there haven't been payouts for an extended period of time.  Previously, the last payout would've been so long ago that any new entrant would've been considered to be in the entire time even though they may have just added.  This would really only be an issue for long epochs but it was still an issue to address.