fix(contracts): Fixed bug in addStake when searching for a stakers existing stake.
addStake has to either find the existing staker, or, if not found, determine the first empty slot to add the staker to.
The existing code was exiting the loop early upon finding the first free slot (likely introduced from early change in logic).
This meant that a staker leaving the pool entirely and in a slot prior to an existing staker, followed by a staker already in the pool adding additional stake, would cause the staker to be in two different slots in the pool - adding in first empty slot instead of incrementing existing stake.

A test was added to ensure proper behavior.