fix(ui): invalidate cached queries on pool/stake mutation
Previously optimistic updates were used to update the cache after a pool or stake mutation. This is cumbersome and prone to errors, since the logic to immutably update deeply-nested data in multiple queries can be complex.

Instead, we now invalidate the cache for queries that return validator, pool, or stake data that was mutated. This is done by calling `invalidateQueries` with a predicate that matches the keys of the affected queries, then calling `router.invalidate()` to trigger refetches.

To avoid refetching all (mostly unaffected) validators, the `['validators']` query is still optimistically updated as before.