feat(ui): simulate donut chart clicks on dropdown change
The only way to highlight a pool in the donut chart is to click it. This simulates that behavior when selecting a pool from the dropdown menu in StakingDetails.

When selecting a pool from the menu, the donut chart will highlight the corresponding pool. It will reset if "All Pools" is selected.

If a staking pool has no stake, it must still appear in the chart for this to work. The solution here is a bit of a hack: in the chart, pools with no stake are set to 1 microalgo. It is displayed as a barely perceptible sliver in the chart, but it is enough to allow the chart to highlight the pool when selected from the dropdown. Its tooltip shows the correct amount of 0 stake.