feat(contracts): stakers can now be removed from a pool by the validator.
The removeStake method in stakingPool.algo.ts has been updated to include a staker parameter.  Previously the txn sender was always the staker.
This change allows the owner or manager of the validator (as sender) to remove the specified staker's stake, so only a staker can remove their stake, or the owner | manager of the validator.
This allows the validator to kick out stakers from a pool, either to refund them if the validator is shutting down their pool, or if they want
(via nodemgr automation perhaps) to kick stakers out of the pool because they no longer meet the gating requirements the validator has defined.