feat: Update contracts, nodemgr, UI to define epochs by number of rounds instead of minutes.
Changed tracking in staking pools from time-based to round-based to better accommodate Algorand's consensus protocol and to allow the node manager and contracts to work better together on when the epoch update is able to be called.  Being round-based also ensures that the +320 round balance lag in online stake is explicitly part of the stakers entry time (simply adding 320 rounds to their entry round).
This commit covers most changes but there are still changes needing to be made in the UI to handle the change in showing Eligible reward %, and for the node manager to no longer use time waits but round-based.