fix(ui): refetch new validator data instead of manually updating (#118)
* fix(ui): refetch new validator data instead of manually updating

Manual updates to the "all validators" query cache are complex, prone to error, and brittle. Refetching the validator after a mutation (e.g., adding a new pool or staking/unstaking) and updating just that validator in the collection with real data is a better solution in every way. The only cost is a single extra network request.

This fixes a bug where `validator.pools` was not being updated after adding a pool or staking/unstaking.

* chore(ui): revert 2532b4e which didn't actually address issue

* refactor(ui): stakesQueryOptions

* chore(ui): rename 'validator-pools' query key to 'pools-info'