test(ui): add tests/docs and fixture data for contracts.ts
This also refactors calculateMaxStake for simplicity and changes its return type to `bigint`, as I've been trying to keep all Algo/asset amounts as BigInts if possible.

ValidatorTable is updated to convert the value to a number before passing it to AlgoAmount. (Hopefully a future version of algokit-utils will update this utility class to support BigInts.)