fix(ui): instantiate KMD client inside dev-only function
The KMD client is only exposed in localnet environment, but it was being instantiated at the top level of /src/utils/devlopment.tsx

This broke the app in staging/prod (i.e., testnet). Now it is instantiated inside the `incrementRoundNumberBy` function, so only when used by the development-only `simulateEpoch` function.