fix(ui): invalidate 'pools-info' query after box value is set (#167)
* Revert "fix(ui): catch and handle 'stakers' box value 404 errors (#165)"

This reverts commit 8cdafb1a9aeea2bca74b0bc0915b165e295fade2.

* fix(ui): wait until step 2 to invalidate 'pools-info' query

In the add pool workflow, the 'pools-info' query was being invalidated after step 1 when the pool is still in the "limbo" state before the storage MBR is paid. That's why fetchStakedInfoForPool was throwing a 404 error: the 'stakers' box value has not been set yet.

This moves invalidation to step 2, after the box value has been set. No more error.