feat(ui): AddValidatorForm entry gating updates
Previously, the `AddValidatorForm` fieldset for entry gating had a simple UI for gating by asset ID. The user would enter 1-4 asset IDs and as long as they were positive integers they would pass validation and be sent to the registry contract on submission.

This updates the asset ID fields to use the `AssetLookup` component created for setting a reward token. It performs a debounced lookup of the input value and displays the asset's unit name if found (failing validation if not).

We've also decided on some additional constraints for setting a minimum balance: it can only be set if there is a single gating asset. Since this field expects a whole unit amount (which will be converted to base units), only one asset's `decimals` param can be used for the conversion.

As soon as a 2nd asset is added, the field is hidden from the form. It is also hidden if the gating type is set to creator address/NFD. In these cases the minimum balance will default to 1 (base unit), effectively checking for any amount of a qualifying asset to allow entry.