fix(contracts): for 'partial epoch' stakers, the reward calculation was incorrect.
The algo reward amount was being reduced as partial epoch stakers were paid out so that full-epoch stakers would use the adjusted amount.  However, the percentage reward against which the partial stakers reward was calculated was this diminishing amount instead of the original total available reward.
Updated one of the 'full and partial stakers' tests to validate w/ 2 partial stakers.